Set expired cards as inactive
